pyproject.toml
[tool.poetry]
name = "social"
description = "Social Media (postuser)"
version = "0.1.0"
readme = "README.md"
authors = ["myapp.com"]
classifiers = [
  "Private :: Do not Upload",
]


[tool.poetry.dependencies]
python = "3.12.1"

django = { version = "^5.0", extras = ["argon2"] }
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-split-settings = "^1.2"
django-axes = "^6.2"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.18"
django-stubs-ext = "^4.2"

psycopg2-binary = "^2.9"
gunicorn = "^21.2"
python-decouple = "^3.8"
structlog = "^23.2"

django-filter = "^23.5"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2"
django-querycount = "^0.8"
django-migration-linter = "^5.0"
django-extra-checks = "^0.13"
nplusone = "^1.0"

wemake-python-styleguide = "^0.18"
flake8-pytest-style = "^1.7"
flake8-django = "^1.4"
flake8-logging-format = "^0.9"
nitpick = "^0.34"
doc8 = "^1.0"

pytest = "^7.4"
pytest-django = "^4.7"
pytest-cov = "^4.0"
covdefaults = "^2.3"
django-coverage-plugin = "^3.1"
pytest-randomly = "^3.15"
pytest-timeout = "^2.2"
django-test-migrations = "^1.3"
hypothesis = "^6.92"

django-stubs = { version = "^4.2.7", extras = ["compatible-mypy"] }

djlint = "^1.34"
yamllint = "^1.33"
safety = "^2.3"
dotenv-linter = "^0.4"
polint = "^0.4"
dennis = "^1.1"
dump-env = "^1.3"
ipython = "^8.18"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2"
sphinx-autodoc-typehints = "^1.25"
tomli = "^2.0"


[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"


[tool.djlint]
ignore = "H006,H030,H031"
include = "H017,H035"
indent = 2
blank_line_after_tag = "load,extends"
profile = "django"
max_line_length = 80
format_attribute_template_tags = true


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/0.18.0/styles/nitpick-style-wemake.toml"
